+++ input_reader.h
@@ -21,6 +21,6 @@ std::string ReadLine();
 int ReadLineWithNumber();
 Query ParseQuery(std::string);
 std::vector<std::string_view> SplitIntoWords(std::string_view);
-std::vector<std::pair<int16_t, std::string>> ParseStopDistances(std::string_view info_);
+std::vector<std::pair<long unsigned int, std::string>> ParseStopDistances(std::string_view info_);
 void CreateBase(TransportCatalogue&);
